Reference

Index

DedekindCutArithmetic.DedekindCutArithmetic

DedekindCutArithmetic.AbstractCut

DedekindCutArithmetic.AbstractDedekindReal

DedekindCutArithmetic.AbstractDyadic

DedekindCutArithmetic.BinaryCompositeCut

DedekindCutArithmetic.DedekindCut

DedekindCutArithmetic.DyadicInterval

DedekindCutArithmetic.DyadicReal

DedekindCutArithmetic.RationalCauchyCut

DedekindCutArithmetic.UnaryCompositeCut

Base.split

DedekindCutArithmetic.dual

DedekindCutArithmetic.exists

DedekindCutArithmetic.forall

DedekindCutArithmetic.high

DedekindCutArithmetic.isbackward

DedekindCutArithmetic.isforward

DedekindCutArithmetic.low

DedekindCutArithmetic.make_prec

DedekindCutArithmetic.midpoint

DedekindCutArithmetic.overlaps

DedekindCutArithmetic.preprocess_expression

DedekindCutArithmetic.radius

DedekindCutArithmetic.refine!

DedekindCutArithmetic.refine!

DedekindCutArithmetic.thirds

DedekindCutArithmetic.width

DedekindCutArithmetic.@cut

DedekindCutArithmetic.@∀

DedekindCutArithmetic.@∃

DedekindCutArithmetic.DedekindCutArithmetic — Module

DedekindCutArithmetic

Julia library implementing exact real arithmetic using Dedekind cuts and abstract stone duality.

DedekindCutArithmetic.AbstractCut — Type

Abstract type for a generic cut.

DedekindCutArithmetic.AbstractDedekindReal — Type

Abstract type representing a generic real number.

DedekindCutArithmetic.AbstractDyadic — Type

Represents a dyadic number or a dyadic interval.

DedekindCutArithmetic.BinaryCompositeCut — Type

Composite cut lazily representing the result of applying an arithmetic binary operation f to child1 and child2.

DedekindCutArithmetic.DedekindCut — Type

Representation of a real number $x$ as a dedekind cut.

Fields

low : function approximating the number from below. Evaluates to true for every number $< x$.

high : function approximating the number from above. Evaluates to true for every number $> x$.

mpa : Cached most precise approximation computed so far. This is a dyadic interval bound $x$ which is updated every time the cut is refined to a higher precision.

DedekindCutArithmetic.DyadicInterval — Type

Represents an interval with lower and uppper bound expressed as dyadic numbers. Note this is a generalized interval, that is, the lower bound can be greater than the upper bound.

DedekindCutArithmetic.DyadicReal — Type

Represents a dyadic number in the form $\frac{m}{2^{-e}}$, with $d ∈ ℤ$ and $e ∈ ℕ$.

DedekindCutArithmetic.RationalCauchyCut — Type

This represents a dyadic interval [(m-1)/2^e, (m+1)/2^e] approximating a non-dyadic rational number.

This is a special-case of Dedekind cut representing a rational number for which we do not need iterative refinement

DedekindCutArithmetic.UnaryCompositeCut — Type

Composite cut lazily representing the result of applying an arithmetic unary operation f to child.

Base.split — Method

Split the interval in two halves.

DedekindCutArithmetic.dual — Method

Given the interval $[a, b]$, return its dual $[b, a]$.

DedekindCutArithmetic.exists — Method

Check whether $∃ x ∈ dom : f(x) < 0$

DedekindCutArithmetic.forall — Method

Check whether $∀ x ∈ dom : f(x) < 0$

DedekindCutArithmetic.high — Method

Lower bound of the real number approximated by the cut

DedekindCutArithmetic.isbackward — Method

Returns true if the lower bound is strictly bigger than the upper bound.

DedekindCutArithmetic.isforward — Method

Returns true if the lower bound is smaller or equal of the upper bound.

DedekindCutArithmetic.low — Method

Lower bound of the real number approximated by the cut

DedekindCutArithmetic.make_prec — Method

Given precision p and interval i, compute a precision which is better than p and is suitable for working with intervals of width i.

Taken from: https://github.com/andrejbauer/marshall/blob/c9f1f6466e879e8db11a12b9bc030e62b07d8bd2/src/eval.ml#L22-L26

DedekindCutArithmetic.midpoint — Method

Midpoint of the current most precise approximation of the cut

DedekindCutArithmetic.overlaps — Method

Whether or not two cuts are overlapping in the current precision

DedekindCutArithmetic.preprocess_expression — Method

Transformations on the expression before being processed by @cut

DedekindCutArithmetic.radius — Method

Width of the current precise appoximation of t mosthe cut

DedekindCutArithmetic.refine! — Method

Refine the given cut to give an approximation with precision bits of accuracy. If the required accuracy cannot be achieved within max_iter iterations, return the current estimate with a warning.

DedekindCutArithmetic.refine! — Method

Evaluate the Cauchy sequence representing a rational number with n bits of precision.

DedekindCutArithmetic.thirds — Method

Return two points that divide the interval in three parts with ratio 1/4, 1/2, 1/2

DedekindCutArithmetic.width — Method

Width of the current precise appoximation of t mosthe cut

DedekindCutArithmetic.@cut — Macro

Macro to construct a DedekindCut.

A cut is defined using the following syntax

@cut x ∈ [a, b], low, high

This defines a real number $x$ in the interval $[a, b]$ which is approximated by the inequality $φ(x)$ from below and $ψ(x)$ from above.

$φ$ and $ψ$ can be any julia expression, also referring to variables in the scope where the macro is called.

Similarly, $a$ and $b$ can also be julia expressions, but they cannot depend on $x$.

DedekindCutArithmetic.@∀ — Macro

Check whether $∀ x ∈ dom : prop(x)$

DedekindCutArithmetic.@∃ — Macro

Check whether $∃ x ∈ dom : prop(x)$